src/video_generation/generator.py
"""
Core video generation module for A250 project.
Handles the generation of 60-second history videos using Hugging Face models.
"""

import os
from dataclasses import dataclass
from pathlib import Path
from typing import Optional, List, Dict

import torch
from huggingface_hub import HfApi
from transformers import AutoModelForTextToVideo, AutoTokenizer
from moviepy.editor import VideoFileClip, TextClip, CompositeVideoClip
from PIL import Image

@dataclass
class VideoConfig:
    """Configuration for video generation."""
    width: int = 1080
    height: int = 1920  # TikTok vertical format
    fps: int = 30
    duration: int = 60  # 60 seconds
    model_name: str = "damo-vilab/text-to-video-ms-1.7b"  # Example model, to be updated based on research
    output_format: str = "mp4"
    temp_dir: str = "temp"

class VideoGenerator:
    """Main class for generating history videos."""
    
    def __init__(self, config: Optional[VideoConfig] = None):
        """Initialize the video generator with configuration."""
        self.config = config or VideoConfig()
        self._setup_directories()
        self._load_models()
        
    def _setup_directories(self):
        """Create necessary directories for video generation."""
        os.makedirs(self.config.temp_dir, exist_ok=True)
        
    def _load_models(self):
        """Load required Hugging Face models."""
        self.tokenizer = AutoTokenizer.from_pretrained(self.config.model_name)
        self.model = AutoModelForTextToVideo.from_pretrained(self.config.model_name)
        if torch.cuda.is_available():
            self.model = self.model.to("cuda")
            
    def generate_from_script(self, script: str, output_path: str) -> str:
        """
        Generate a video from a script.
        
        Args:
            script: The script text to generate video from
            output_path: Where to save the generated video
            
        Returns:
            Path to the generated video
        """
        # Tokenize the script
        inputs = self.tokenizer(script, return_tensors="pt")
        if torch.cuda.is_available():
            inputs = {k: v.to("cuda") for k, v in inputs.items()}
            
        # Generate video frames
        with torch.no_grad():
            video_frames = self.model.generate(**inputs)
            
        # Convert frames to video
        video_path = self._frames_to_video(video_frames, output_path)
        
        # Add text overlays and effects
        final_video = self._post_process_video(video_path)
        
        return final_video
        
    def _frames_to_video(self, frames: torch.Tensor, output_path: str) -> str:
        """Convert generated frames to a video file."""
        # Implementation will depend on the specific model output format
        # This is a placeholder for the actual implementation
        pass
        
    def _post_process_video(self, video_path: str) -> str:
        """Add text overlays, effects, and optimize for TikTok."""
        video = VideoFileClip(video_path)
        
        # Add text overlays
        # Add effects
        # Optimize for TikTok format
        
        return video_path
        
    def cleanup(self):
        """Clean up temporary files and resources."""
        # Implementation for cleanup
        pass

def generate_video(script_path: str, output_path: str) -> str:
    """
    Convenience function to generate a video from a script file.
    
    Args:
        script_path: Path to the script file
        output_path: Where to save the generated video
        
    Returns:
        Path to the generated video
    """
    with open(script_path, 'r') as f:
        script = f.read()
        
    generator = VideoGenerator()
    try:
        video_path = generator.generate_from_script(script, output_path)
        return video_path
    finally:
        generator.cleanup() 