requirements.txt
# Core dependencies
huggingface-hub>=0.19.0
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.0
pillow>=10.0.0

# Video processing
moviepy>=1.0.3
opencv-python>=4.8.0

# API and web
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0 